# backend/storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMS Project API",
        "description": "    \n    ⚠️ Authorization Guide:\n    1. Use API **Login** to get Access Token.\n    2. Click on **[Authorize]** button.\n    3. Paste the token into the field.\n    4. If you receive **401**, it means you haven't logged in or the token has expired.\n    5. If you receive **403**, it means you don't have permission to access.\n\n    *Note: \n    - If you haven't logged in but the system still run the API (due to browser saving cookie), try again with anonymous.\n    - The System verify the token through cookie.\n    ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get permission list",
                "description": "Display a listing of the resource.",
                "operationId": "b4182452579c4e425259845595e4008c",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "currentPage",
                        "in": "query",
                        "description": "Current page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/PermissionsSearch"
                    },
                    "400": {
                        "$ref": "#/components/responses/Exception400"
                    },
                    "401": {
                        "$ref": "#/components/responses/Exception401"
                    },
                    "404": {
                        "$ref": "#/components/responses/Exception404"
                    },
                    "500": {
                        "$ref": "#/components/responses/Exception500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Create permission",
                "description": "Store a newly created resource in storage.",
                "operationId": "79eae836852641979b35824ad81c61b1",
                "requestBody": {
                    "$ref": "#/components/requestBodies/PermissionsCreate"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/PermissionsCreate"
                    },
                    "400": {
                        "$ref": "#/components/responses/Exception400"
                    },
                    "401": {
                        "$ref": "#/components/responses/Exception401"
                    },
                    "404": {
                        "$ref": "#/components/responses/Exception404"
                    },
                    "500": {
                        "$ref": "#/components/responses/Exception500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/permissions/{id}": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get permission by id",
                "description": "Display the specified resource.",
                "operationId": "6b1c14a2af1c3d76f54be0a4c8464097",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Permission id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/PermissionsGetById"
                    },
                    "400": {
                        "$ref": "#/components/responses/Exception400"
                    },
                    "401": {
                        "$ref": "#/components/responses/Exception401"
                    },
                    "404": {
                        "$ref": "#/components/responses/Exception404"
                    },
                    "500": {
                        "$ref": "#/components/responses/Exception500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Update permission",
                "description": "Update the specified resource in storage.",
                "operationId": "1fe296fe2b49effc7fb8aed97aed47b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Permission id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "123456"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/PermissionsUpdate"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/PermissionsUpdate"
                    },
                    "400": {
                        "$ref": "#/components/responses/Exception400"
                    },
                    "401": {
                        "$ref": "#/components/responses/Exception401"
                    },
                    "404": {
                        "$ref": "#/components/responses/Exception404"
                    },
                    "500": {
                        "$ref": "#/components/responses/Exception500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Delete permission",
                "description": "Remove the specified resource from storage.",
                "operationId": "ee6486510642517d73456adfed700f14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Permission id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "123456"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/PermissionsDelete"
                    },
                    "400": {
                        "$ref": "#/components/responses/Exception400"
                    },
                    "401": {
                        "$ref": "#/components/responses/Exception401"
                    },
                    "404": {
                        "$ref": "#/components/responses/Exception404"
                    },
                    "500": {
                        "$ref": "#/components/responses/Exception500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Exception": {
                "required": [
                    "errorCode",
                    "errorMessage",
                    "data"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "ERROR_CODE"
                    },
                    "message": {
                        "type": "string",
                        "example": "Error message."
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "example": []
                    }
                },
                "type": "object"
            },
            "Permissions": {
                "required": [
                    "permission_name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "uniqueItems": true,
                        "example": "admin"
                    },
                    "description": {
                        "type": "string",
                        "example": "web"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2023-01-01 00:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2023-01-01 00:00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "Exception400": {
                "description": "Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Exception401": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Exception404": {
                "description": "Not Found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Exception500": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "PermissionsCreate": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "PermissionsDelete": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "PermissionsGetById": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Permissions"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "PermissionsSearch": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Permissions"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "PermissionsUpdate": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Permissions"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "requestBodies": {
...
            "PermissionsCreate": {
                "description": "Create permission",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "display_name"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "CREATE_PERMISSION"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Create Permission"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "PermissionsUpdate": {
                "description": "Update permission",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "display_name"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "CREATE_PERMISSION"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Create Permission"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}